[build|refactor|test] tests for module library (#62)
#46
