+++ src/module/core.cppm
@@ -8,20 +8,12 @@ module;
 
 #include <algorithm>
 #include <array>
-#include <concepts>
-#include <coroutine>
-#include <exception>
-#include <functional>
-#include <optional>
 #include <ranges>
 #include <source_location>
-#include <span>
 #include <string>
 #include <string_view>
 #include <tuple>
 #include <type_traits>
-#include <variant>
-#include <vector>
 
 export module pbdt;
