+++ tests/include/fixtures.hpp
@@ -8,87 +8,7 @@
 
 #include <array>
 #include <print>
-#include <ranges>
-#include <tuple>
-
-#include "pbdt/test-context.hpp"
-
-namespace natural_components
-{
-    namespace target
-    {
-        constexpr int freeFuncTarget(int x)
-        {
-            return x + 0;
-        }
-
-        constexpr auto funcObjTarget = [](int x) -> int
-        {
-            return freeFuncTarget(x);
-        };
-    }
-
-    namespace prop
-    {
-        constexpr auto freeFuncProp(int x, int result)
-        {
-            return pbdt::test_context::expect(x == result);
-        }
-
-        constexpr auto funcObjProp = [](int x, int result)
-        {
-            return freeFuncProp(x, result);
-        };
-    }
-
-    namespace domain
-    {
-        constexpr std::array<int, 5> containerDomain{
-            1, 2, 3, 4, 5,
-        };
-
-        // generator domain
-
-        constexpr auto viewDomain = containerDomain
-                                  | std::views::transform(
-                                        [](int x)
-                                        {
-                                            return x;
-                                        }
-                                  );
-    }
-
-    namespace target2
-    {
-        constexpr int freeFuncTarget(int x, int y)
-        {
-            return (x + y) - (y + x);
-        }
-    }
-
-    namespace prop2
-    {
-        constexpr auto freeFuncProp(std::tuple<int, int> args, int result)
-        {
-            return pbdt::test_context::expect(result == 0);
-        }
-    }
-
-    namespace domain2
-    {
-        constexpr std::array<int, 5> containerDomain{
-            1, 2, 3, 4, 5,
-        };
-
-        constexpr auto viewDomain = containerDomain
-                                  | std::views::transform(
-                                        [](int x)
-                                        {
-                                            return std::tuple{ x };
-                                        }
-                                  );
-    }
-}
+#include <source_location>
 
 namespace
 {
